+++ inmobiliarias/usuarios/models.py
@@ -15,7 +15,7 @@ class User(AbstractUser):
 class administrador(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, related_name='administrador')
     tipo_documento = models.CharField(max_length=3, choices=TipoDocumento.choices, blank=True, null=True)  # Permitir que este campo esté vacío
-    doc_identificacion = models.IntegerField(blank=True, null=True)  # Permitir que este campo esté vacío
+    doc_identificacion = models.PositiveBigIntegerField(blank=True, null=True)  # Permitir que este campo esté vacío
     lugarExpCedula = models.CharField(blank=True,max_length=255)
     CodClasificaIndustrialIU=models.IntegerField(blank=True, null=True)
     descActividadEconomica = models.CharField(blank=True,max_length=255)
@@ -44,7 +44,7 @@ class administrador(models.Model):
 class propietario(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, related_name='propietario')
     tipo_documento = models.CharField(max_length=3, choices=TipoDocumento.choices, blank=True, null=True)
-    doc_identificacion = models.IntegerField(blank=True, null=True)
+    doc_identificacion = models.PositiveBigIntegerField(blank=True, null=True)
     lugarExpCedula = models.CharField(blank=True,null=True, max_length=255)
     descActividadEconomica = models.CharField(blank=True,null=True,max_length=255)
     CodClasificaIndustrialIU=models.IntegerField(blank=True, null=True)
@@ -73,7 +73,7 @@ class propietario(models.Model):
 class arrendatario(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, related_name='arrendatario')
     tipo_documento = models.CharField(max_length=3, choices=TipoDocumento.choices, blank=True, null=True)
-    doc_identificacion = models.IntegerField(blank=True, null=True)
+    doc_identificacion = models.PositiveBigIntegerField(blank=True, null=True)
     lugarExpCedula = models.CharField(blank=True,null=True,max_length=255)
     ocupacion = models.CharField(blank=True,null=True,max_length=255)
     estadoCivil = models.CharField(blank=True,null=True,max_length=255)
@@ -89,8 +89,8 @@ class arrendatario(models.Model):
     ciudadCorrespondencia = models.CharField(blank=True,max_length=255) 
     activo= models.BooleanField(default=True)
     arrendatarioAprobado= models.BooleanField(default=False)
-    celular = models.IntegerField(blank=True, null=True)
-    celularDos = models.IntegerField(blank=True, null=True) 
+    celular = models.PositiveBigIntegerField(blank=True, null=True)
+    celularDos = models.PositiveBigIntegerField(blank=True, null=True) 
 
 
     def __str__(self):
@@ -104,7 +104,7 @@ class arrendatario(models.Model):
 class proveedor(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE, related_name='proveedor')
     tipo_documento = models.CharField(max_length=3, choices=TipoDocumento.choices, blank=True, null=True)
-    doc_identificacion = models.IntegerField(blank=True, null=True)
+    doc_identificacion = models.PositiveBigIntegerField(blank=True, null=True)
     lugarExpCedula = models.CharField(blank=True,null=True,max_length=255)
     genero = models.CharField(blank=True,null=True,max_length=1,choices=Genero.choices)
     direccion = models.CharField(blank=True,null=True,max_length=255)
@@ -114,8 +114,8 @@ class proveedor(models.Model):
     barrioCorrespondencia = models.CharField(blank=True,max_length=255)
     ciudadCorrespondencia = models.CharField(blank=True,max_length=255)     
     activo= models.BooleanField(default=True)
-    celular = models.IntegerField(blank=True, null=True)
-    celularDos = models.IntegerField(blank=True, null=True) 
+    celular = models.PositiveBigIntegerField(blank=True, null=True)
+    celularDos = models.PositiveBigIntegerField(blank=True, null=True) 
 
     def __str__(self):
         return self.user.username
